Sun, Jan 23, 2022, 2:13 PM +00:00

+++ CS50x/week1/pset1/credit/tester.c
@@ -9,7 +9,7 @@ int getIntLength();
 
 int main(void) {
     long long i = get_long_long("Gimme a long: ");
-    int len = getIntLength(i);
+    long len = getIntLength(i);
     printf("len is: %i\n", len);
     // Converting Int to String
